backend/requirements.txt
antlr4-python3-runtime==4.12.0
asgiref==3.8.1
bitarray==3.0.0
bitstring==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
holidays==0.69
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
kazoo==2.10.0
MarkupSafe==3.0.2
oauthlib==3.2.2
overrides==7.7.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
pyutils==0.0.14
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.5.6
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
